Separated private message model
